no-jira: DRY up Helpers
